#88 - 유저 회원 탈퇴 시 운영 중인 도전방이 있는 경우 400 에러 반환 (ROOM_DELETE_REQUIRED)

+++ src/main/java/com/dodal/meet/service/UserService.java
@@ -257,6 +257,12 @@ public class UserService {
     @Transactional
     public void deleteUser(Authentication authentication) {
         UserEntity userEntity = userToUserEntity(authentication);
+
+        List<ChallengeRoomEntity> hostRoomList = challengeRoomEntityRepository.findAllByHostId(userEntity.getId());
+        if (!CollectionUtils.isEmpty(hostRoomList)) {
+            throw new DodalApplicationException(ErrorCode.ROOM_DELETE_REQUIRED);
+        }
+
         userEntityRepository.delete(userEntity);
     }
